requirements.txt
argparse
PyPDF2
atlassian-python-api